+++ include/editor/HierarchyEditor.hpp
@@ -7,7 +7,7 @@ namespace Guarneri
 	class HierarchyEditor : public BaseEditor
 	{
 	public:
-		HierarchyEditor(float x, float y, float w, float h);
+		HierarchyEditor(int x, int y, int w, int h);
 		virtual void on_gui();
 		void draw_transform(Transform* transform);
 		void draw_scene_hierarchy();
